@@ -3,24 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.10.4",
-    "@mui/material": "^5.0.3",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.5.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
-    "react-github-calendar": "^2.0.2",
-    "react-icons": "^4.2.0",
-    "react-parallax-tilt": "^1.4.68",
-    "react-particles-js": "^3.4.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "typewriter-effect": "^2.17.0",
-    "web-vitals": "^0.2.4"
+    "@emotion/react": "^11.13.5",
+    "@mui/material": "^5.16.10",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@tsparticles/engine": "^3.8.0",
+    "@tsparticles/react": "^3.0.0",
+    "@tsparticles/slim": "^3.8.0",
+    "ajv": "6.12.6",
+    "ajv-formats": "1.6.1",
+    "ajv-keywords": "3.5.2",
+    "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.6",
+    "react-dom": "^18.3.1",
+    "react-github-calendar": "^4.4.1",
+    "react-icons": "^5.4.0",
+    "react-parallax-tilt": "^1.7.267",
+    "react-router-dom": "^6.28.0",
+    "react-scripts": "5.0.1",
+    "typewriter-effect": "^2.21.0",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
